// src/shared/enums/duration.enum.ts
export enum DurationTimeEnum {
  none_duration = 'none',
  fifteen_minute_duration = '15',
  thirty_minute_duration = '30',
  forty_five_duration = '45',
  one_hour_duration = '60',
  two_hour_duration = '120',
  three_hour_duration = '180',
  four_hour_duration = '240',
  five_hour_duration = '300',
  six_hour_duration = '360',
  seven_hour_duration = '420',
}